requirements.txt
django
pre-commit
djangorestframework
django-filter
djangorestframework-api-key
